Adds annotated image capture ability, and refactored image annotation logic into a widget
